gh-1644 Add AssetsViewController for holding total balance header and coins & collectibles tabs
