+++ Multisig/UI/Assets/AssetsViewController.xib
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19455" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19454"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AssetsViewController" customModule="Multisig" customModuleProvider="target">
+            <connections>
+                <outlet property="contentView" destination="jwx-DX-oEe" id="6EI-Ew-daR"/>
+                <outlet property="totalBalanceView" destination="JaT-py-F36" id="IlT-hu-QFd"/>
+                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="JaT-py-F36" customClass="TotalBalanceView" customModule="Multisig" customModuleProvider="target">
+                    <rect key="frame" x="0.0" y="44" width="414" height="150"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="150" id="cX1-Ag-6xP"/>
+                    </constraints>
+                </view>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jwx-DX-oEe">
+                    <rect key="frame" x="0.0" y="194" width="414" height="668"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                </view>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="jwx-DX-oEe" firstAttribute="top" secondItem="JaT-py-F36" secondAttribute="bottom" id="5QA-dT-Qu2"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="JaT-py-F36" secondAttribute="trailing" id="Fnb-Av-Kes"/>
+                <constraint firstItem="JaT-py-F36" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="LmH-Uc-QAE"/>
+                <constraint firstItem="JaT-py-F36" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="V2F-ct-UK9"/>
+                <constraint firstItem="jwx-DX-oEe" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="YwL-HB-5pg"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="jwx-DX-oEe" secondAttribute="bottom" id="ej1-tJ-aV3"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="jwx-DX-oEe" secondAttribute="trailing" id="oFL-gB-bB9"/>
+            </constraints>
+            <point key="canvasLocation" x="132" y="131"/>
+        </view>
+    </objects>
+    <resources>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
